fix(spec): rendering a Pristine document should ensure raw is up to date
This PR makes sure that whenever rendering a Pristine document (e.g. to
create a clean document from a fresh clone), the raw root doc is
up-to-date.

In go-swagger, we modify the spec (e.g. with flatten, expand, ...) and
the internal state of the spec document may sometimes need a clean
Document to be reconstructed.

* contributes go-swagger/go-swagger#2346

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>